refactor const
